Refactored sockets to notification service
