+++ server/server.js
@@ -6,7 +6,8 @@
 var init = require('./config/init')(),
     config = require('./config/config'),
     mongoose = require('mongoose-q')(),
-    path = require('path');
+    path = require('path'),
+    socket = require('./app/socket/main.socket');
 
 
 // Bootstrap db connection
@@ -17,15 +18,17 @@ var app = require('./config/express')(db);
 
 // Init socket.io
 var server = require('http').Server(app);
-var io =  require('socket.io')(server);
-//expose socket.io
-exports.io = io;
-require('./app/socket/main.socket')();
+var socketIo = require('socket.io')(server);
+socket.init(socketIo);
+
+// Load notification service after sockets and models
+require('./app/services/notification.service')
 
 //require all routes
 config.getGlobbedFiles('./app/routes/**/*.js').forEach(function(routePath) {
     require(path.resolve(routePath))(app);
 });
+
 
 // Start the app by listening on <port>
 server.listen(config.port);
